feat: 🎸 Add api management directory
✅ Closes: #145
