+++ src/app.module.ts
@@ -19,37 +19,37 @@
  ********************************************************************************/
 
 import { Module } from "@nestjs/common"
-import { VersionController } from "./version/version.controller"
-import { VersionService } from "./version/version.service"
-import { SyncController } from "./sync/sync.controller"
-import { SyncService } from "./sync/sync.service"
-import { TipsetController } from "./tipset/tipset.controller"
-import { TipsetService } from "./tipset/tipset.service"
-import { BlockController } from "./block/block.controller"
-import { BlockService } from "./block/block.service"
-import { MessageController } from "./message/message.controller"
-import { MessageService } from "./message/message.service"
-import { DataswapMessageController } from "./dataswapMessage/dataswapMessage.controller"
-import { DataswapMessageService } from "./dataswapMessage/dataswapMessage.service"
-import { DatasetMetadataController } from "./datasetMetadata/datasetMetadata.controller"
-import { DatasetMetadataService } from "./datasetMetadata/datasetMetadata.service"
-import { CarController } from "./car/car.controller"
-import { CarService } from "./car/car.service"
-import { CarReplicaController } from "./carReplica/carReplica.controller"
-import { CarReplicaService } from "./carReplica/carReplica.service"
-import { DatasetRequirementController } from "./datasetRequirement/datasetRequirement.controller"
-import { DatasetRequirementService } from "./datasetRequirement/datasetRequirement.service"
-import { DatasetProofMetadataController } from "./datasetProofMetadata/datasetProofMetadata.controller"
-import { DatasetProofMetadataService } from "./datasetProofMetadata/datasetProofMetadata.service"
-import { MatchingMetadataController } from "./matchingMetadata/matchingMetadata.controller"
-import { MatchingMetadataService } from "./matchingMetadata/matchingMetadata.service"
-import { MatchingTargetController } from "./matchingTarget/matchingTarget.controller"
-import { MatchingTargetService } from "./matchingTarget/matchingTarget.service"
-import { MatchingBidsController } from "./matchingBids/matchingBids.controller"
-import { MatchingBidsService } from "./matchingBids/matchingBids.service"
+import { VersionController } from "./api/version/version.controller"
+import { VersionService } from "./api/version/version.service"
+import { SyncController } from "./api/sync/sync.controller"
+import { SyncService } from "./api/sync/sync.service"
+import { TipsetController } from "./api/tipset/tipset.controller"
+import { TipsetService } from "./api/tipset/tipset.service"
+import { BlockController } from "./api/block/block.controller"
+import { BlockService } from "./api/block/block.service"
+import { MessageController } from "./api/message/message.controller"
+import { MessageService } from "./api/message/message.service"
+import { DataswapMessageController } from "./api/dataswapMessage/dataswapMessage.controller"
+import { DataswapMessageService } from "./api/dataswapMessage/dataswapMessage.service"
+import { DatasetMetadataController } from "./api/datasetMetadata/datasetMetadata.controller"
+import { DatasetMetadataService } from "./api/datasetMetadata/datasetMetadata.service"
+import { CarController } from "./api/car/car.controller"
+import { CarService } from "./api/car/car.service"
+import { CarReplicaController } from "./api/carReplica/carReplica.controller"
+import { CarReplicaService } from "./api/carReplica/carReplica.service"
+import { DatasetRequirementController } from "./api/datasetRequirement/datasetRequirement.controller"
+import { DatasetRequirementService } from "./api/datasetRequirement/datasetRequirement.service"
+import { DatasetProofMetadataController } from "./api/datasetProofMetadata/datasetProofMetadata.controller"
+import { DatasetProofMetadataService } from "./api/datasetProofMetadata/datasetProofMetadata.service"
+import { MatchingMetadataController } from "./api/matchingMetadata/matchingMetadata.controller"
+import { MatchingMetadataService } from "./api/matchingMetadata/matchingMetadata.service"
+import { MatchingTargetController } from "./api/matchingTarget/matchingTarget.controller"
+import { MatchingTargetService } from "./api/matchingTarget/matchingTarget.service"
+import { MatchingBidsController } from "./api/matchingBids/matchingBids.controller"
+import { MatchingBidsService } from "./api/matchingBids/matchingBids.service"
 import { BackgroundTaskService } from "./backgroundTask/provider/backgroundTask.service"
-import { DatasetChallengeController } from "./datasetChallenge/datasetChallenge.controller"
-import { DatasetChallengeService } from "./datasetChallenge/datasetChallenge.service"
+import { DatasetChallengeController } from "./api/datasetChallenge/datasetChallenge.controller"
+import { DatasetChallengeService } from "./api/datasetChallenge/datasetChallenge.service"
 
 /**
  * Root module for the application.
